pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Show difference between two data files"
authors = ["Georgy <nov_georgy@inbox.ru>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
generate-diff = "gendiff.scripts.generate_diff:main"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
